+++ src/lib.rs
@@ -27,6 +27,7 @@ pub extern crate libc;
 
 pub use errno::Errno;
 
+pub mod dirent;
 pub mod errno;
 pub mod features;
 pub mod fcntl;